Updated nav bar to be custom to user type
Nav bar displays differently if you are not logged in vs. logged in as a
student vs. logged in as a professor.
For a student, the "Grades" link from account home goes to all grades
for all classes.  The "Grades" link when on a class page only pulls the
grades for that specific class.  Need to update this page to show
assignment averages, low and high,...etc.